+++ src/main/java/baseball/util/InputUtil.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.util;
 
 import camp.nextstep.edu.missionutils.Console;
